Moved ApplicantListItem to its own file as a sub-component of ApartmentApplicantList
