+++ src/components/ApartmentApplicantList/apartmentApplicantList.scss
@@ -1,12 +1,5 @@
 @import '../../vars';
 
-// TODO - deprecated. delete this file
-
 .apartment-applicant-list {
   border-radius: 4px;
-  .applicant-list-header {
-    border-radius: 4px;
-    color: $neutral-white;
-    background-color: $primary-blue;
-  }
 }
