+++ src/components/ApartmentApplicantList/components/ApplicantListItem/applicantListItem.scss
@@ -1,0 +1,7 @@
+@import '../../vars';
+
+.applicant-list-header {
+  border-radius: 4px;
+  color: $neutral-white;
+  background-color: $primary-blue;
+}
